Code cleanup
-Removed scanf/sscanf compiler warrnings
-Fixed some integer casting to avoid compiler warnings
-Do range check on starting private key value
-Fixed commandline processing in AddrGen.exe

+++ util/util.cpp
@@ -54,7 +54,7 @@ namespace util {
 
 		std::string s(buf);
 
-		int len = s.length();
+		int len = (int)s.length();
 
 		int numCommas = (len - 1) / 3;
